refactor(py): clean public imports in samples and remove outdated fastapi-bugbot and middleware-coding-agent samples
