+++ py/pyproject.toml
@@ -142,13 +142,11 @@ context              = { workspace = true }
 django-hello         = { workspace = true }
 dynamic-tools        = { workspace = true }
 evaluators           = { workspace = true }
-fastapi-bugbot       = { workspace = true }
 flask-hello          = { workspace = true }
 gemini-code-execution = { workspace = true }
 gemini-context-caching = { workspace = true }
 google-genai-media   = { workspace = true }
 middleware           = { workspace = true }
-middleware-coding-agent = { workspace = true }
 ollama-sample        = { workspace = true }
 output-formats       = { workspace = true }
 prompts              = { workspace = true }
