+++ py/samples/django-hello/recipes/views.py
@@ -24,8 +24,7 @@ from genkit_django import genkit_django_handler
 from genkit_google_genai import GoogleAI
 from pydantic import BaseModel, Field
 
-from genkit import Genkit, ModelResponse
-from genkit._core._action import ActionRunContext
+from genkit import ActionRunContext, Genkit, ModelResponse
 from genkit.plugin_api import RequestData
 
 ai = Genkit(
