--- py/samples/fastapi-bugbot/prompts/analyze_diff.prompt
@@ -1,27 +1,0 @@
----
-model: googleai/gemini-flash-latest
-input:
-  schema:
-    diff: string
-    context?: string
-output:
-  schema: Analysis
----
-
-You are a code reviewer. Analyze this code diff for issues.
-
-{{#if context}}
-Context: {{context}}
-{{/if}}
-
-Diff:
-```diff
-{{diff}}
-```
-
-Focus on changes that introduce:
-- Security vulnerabilities
-- Bugs or logic errors
-- Style violations
-
-Return ONLY issues in the changed lines.
