--- py/samples/fastapi-bugbot/prompts/analyze_style.prompt
@@ -1,24 +1,0 @@
----
-model: googleai/gemini-flash-latest
-input:
-  schema:
-    code: string
-    language?: string
-output:
-  schema: Analysis
----
-
-You are a code style expert. Analyze the following {{language}} code for style issues and best practices.
-
-Code:
-```{{language}}
-{{code}}
-```
-
-Focus on:
-- Naming conventions
-- Code organization
-- Documentation
-- Idiomatic patterns for {{language}}
-
-Return ONLY important style issues.
