--- py/samples/fastapi-bugbot/pyproject.toml
@@ -1,18 +1,0 @@
-[project]
-name = "fastapi-bugbot"
-version = "0.2.0"
-requires-python = ">=3.10"
-dependencies = [
-  "genkit",
-  "genkit-fastapi",
-  "genkit-google-genai",
-  "python-dotenv>=1.0.0",
-  "uvicorn[standard]>=0.34.0",
-]
-
-[build-system]
-build-backend = "hatchling.build"
-requires = ["hatchling"]
-
-[tool.hatch.build.targets.wheel]
-packages = ["src"]
