+++ py/samples/flask-hello/src/main.py
@@ -23,9 +23,8 @@ from genkit_flask import genkit_flask_handler
 from genkit_google_genai import GoogleAI
 from pydantic import BaseModel, Field
 
-from genkit import Genkit, ModelResponse
-from genkit._core._action import ActionRunContext
-from genkit._core._context import RequestData
+from genkit import ActionRunContext, Genkit, ModelResponse
+from genkit.plugin_api import RequestData
 
 ai = Genkit(
     plugins=[GoogleAI()],
