--- py/samples/middleware-coding-agent/.gitignore
@@ -1,4 +1,0 @@
-__pycache__/
-
-# The agent edits files here in place; outputs are reproducible by re-running.
-workspace/
