--- py/samples/middleware-coding-agent/pyproject.toml
@@ -1,18 +1,0 @@
-[project]
-name = "middleware-coding-agent"
-version = "0.1.0"
-requires-python = ">=3.10"
-dependencies = [
-  "genkit",
-  "genkit-google-genai",
-  "genkit-middleware",
-  "pydantic>=2.10.5",
-  "structlog>=25.2.0",
-]
-
-[build-system]
-build-backend = "hatchling.build"
-requires = ["hatchling"]
-
-[tool.hatch.build.targets.wheel]
-packages = ["src"]
